+++ src/snowflake/cli/_plugins/spcs/services/commands.py
@@ -47,7 +47,6 @@ from snowflake.cli.api.exceptions import (
     IncompatibleParametersError,
     NoProjectDefinitionError,
 )
-from snowflake.cli.api.feature_flags import FeatureFlag
 from snowflake.cli.api.identifiers import FQN
 from snowflake.cli.api.output.types import (
     CollectionResult,
@@ -361,7 +360,6 @@ def logs(
 
 @app.command(
     requires_connection=True,
-    is_enabled=FeatureFlag.ENABLE_SPCS_SERVICE_EVENTS.is_enabled,
 )
 def events(
     name: FQN = ServiceNameArgument,
@@ -409,7 +407,6 @@ def events(
 
 @app.command(
     requires_connection=True,
-    is_enabled=FeatureFlag.ENABLE_SPCS_SERVICE_METRICS.is_enabled,
 )
 def metrics(
     name: FQN = ServiceNameArgument,