+++ src/snowflake/cli/api/project/schemas/entities/entities.py
@@ -24,6 +24,8 @@ from snowflake.cli._plugins.nativeapp.entities.application_package import (
     ApplicationPackageEntity,
     ApplicationPackageEntityModel,
 )
+from snowflake.cli._plugins.notebook.notebook_entity import NotebookEntity
+from snowflake.cli._plugins.notebook.notebook_entity_model import NotebookEntityModel
 from snowflake.cli._plugins.snowpark.snowpark_entity import (
     FunctionEntity,
     ProcedureEntity,
@@ -60,6 +62,7 @@ Entity = Union[
     ComputePoolEntity,
     ImageRepositoryEntity,
     ServiceEntity,
+    NotebookEntity,
 ]
 EntityModel = Union[
     ApplicationEntityModel,
@@ -70,6 +73,7 @@ EntityModel = Union[
     ComputePoolEntityModel,
     ImageRepositoryEntityModel,
     ServiceEntityModel,
+    NotebookEntityModel,
 ]
 
 ALL_ENTITIES: List[Entity] = [*get_args(Entity)]